Cache, listAllKnownDomains: Move to cache.go
The function has no correlation to domain watcher.
